requirements.txt
Flask==3.1.1
gunicorn==23.0.0
tensorflow==2.19.0
numpy==2.1.3
pillow==11.2.1
opencv-python-headless==4.11.0
